feat: initial commit for Spin fileserver
This is the first commit for the Spin HTTP fileserver, a Spin HTTP
component that can be used to serve static files in Spin applications.

It is designed to be a drop-in replacement for fileserver written
in Grain and used to currently serve fermyon.com with Wagi.
As such, it supports setting the desired cache header (currently only
a global cache, and not per file type, which can be added in the future),
as well as prefixing the path for resolving files locally.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>