fix: update file path from request URI to PATH_INFO
This commit updates the source of the requested file from request.uri
to the value of the PATH_INFO header.
This is because Spin now populates the URI field on a request according
to the HTTP spec — to include the absolute path of the request, which
could include an application base path and component route.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>